+++ src/ng-vcl/icogram/index.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { VCLLayoutModule } from '../core/index';
 import { VCLIconModule } from '../icon/index';
-import { IcogramComponent } from './icogram.component';
+import { IcogramComponent, IcogramSpanComponent } from './icogram.component';
 
-export { IcogramComponent };
+export { IcogramComponent, IcogramSpanComponent };
 
 @NgModule({
-  imports: [CommonModule, VCLIconModule],
-  exports: [IcogramComponent],
-  declarations: [IcogramComponent],
+  imports: [CommonModule, VCLIconModule, VCLLayoutModule],
+  exports: [IcogramComponent, IcogramSpanComponent, VCLLayoutModule],
+  declarations: [IcogramComponent, IcogramSpanComponent],
   providers: [],
 })
 export class VCLIcogramModule { }